+++ glean-core/benchmark/benches/lifetime_buffering.rs
@@ -31,6 +31,8 @@ pub fn delay_io_benchmark(c: &mut Criterion) {
             ping_schedule: Default::default(),
             ping_lifetime_threshold: 0,
             ping_lifetime_max_time: 0,
+            max_pending_pings_count: None,
+            max_pending_pings_directory_size: None,
         };
         let glean = Glean::new(cfg).unwrap();
 
@@ -73,6 +75,8 @@ pub fn delay_io_benchmark(c: &mut Criterion) {
             ping_schedule: Default::default(),
             ping_lifetime_threshold: 0,
             ping_lifetime_max_time: 0,
+            max_pending_pings_count: None,
+            max_pending_pings_directory_size: None,
         };
         let glean = Glean::new(cfg).unwrap();