+++ glean-core/tests/common/mod.rs
@@ -69,6 +69,8 @@ pub fn new_glean_with_upload(
         ping_schedule: Default::default(),
         ping_lifetime_threshold: 0,
         ping_lifetime_max_time: 0,
+        max_pending_pings_count: None,
+        max_pending_pings_directory_size: None,
     };
     let mut glean = Glean::new(cfg).unwrap();